% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{AnnDatR}
\alias{AnnDatR}
\title{AnnDatR Class}
\description{
A flexible, Anndata-like container for transcriptomics data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{X}}{Primary expression data matrix (features x samples).}

\item{\code{obs}}{Cell/Sample metadata (data.frame).}

\item{\code{var}}{Feature/Gene metadata (data.frame).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnDatR-new}{\code{AnnDatR$new()}}
\item \href{#method-AnnDatR-print}{\code{AnnDatR$print()}}
\item \href{#method-AnnDatR-validate}{\code{AnnDatR$validate()}}
\item \href{#method-AnnDatR-filter_obs}{\code{AnnDatR$filter_obs()}}
\item \href{#method-AnnDatR-filter_var}{\code{AnnDatR$filter_var()}}
\item \href{#method-AnnDatR-clone}{\code{AnnDatR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDatR-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDatR-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDatR$new(
  prefix_name = NULL,
  var_names = "ensembl_ids",
  layer = NULL,
  file_dir = "./",
  X = NULL,
  var = NULL,
  obs = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDatR-print"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDatR-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDatR$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDatR-validate"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDatR-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDatR$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDatR-filter_obs"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDatR-filter_obs}{}}}
\subsection{Method \code{filter_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDatR$filter_obs(column, string, negative = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDatR-filter_var"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDatR-filter_var}{}}}
\subsection{Method \code{filter_var()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDatR$filter_var(column, string, negative = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDatR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDatR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDatR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
